Conexión: Front-end | Back-end | Base de datos

+++ src/main/java/org/project/project/repository/DocumentacionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DocumentacionRepository extends JpaRepository<Documentacion, Integer> {
+public interface DocumentacionRepository extends JpaRepository<Documentacion, Long> {
 }
